Added the abstract Transformer class that all the transforms extend,
as well as changed Queue to contain doubles

added DFT, Patricks' got concurrent dft later, fft later this week

yay

+++ src/Song.java
@@ -1,7 +1,9 @@
+import java.io.File;
+
 public class Song
 {
 	private final int[] data;
-	private final float SAMPLING_FREQUENCY;
+	private final int SAMPLING_FREQUENCY;
 	
 	public Song(File file)
 	{
@@ -14,7 +16,7 @@ public class Song
 		return data;
 	}
 
-	public float getSamplingFreq()
+	public int getSamplingFreq()
 	{
 		return SAMPLING_FREQUENCY;
 	}
